requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pillow==11.0.0
python-decouple==3.8
sqlparse==0.5.1
